refactor: centralize shared assets
