+++ backend/src/modules/users/tutorials/tutorial.service.js
@@ -5,7 +5,7 @@ const chapterService = require("./chapters/tutorialChapter.service");
 const { withTransaction } = require("../../../services/transaction.service");
 const { v4: uuidv4 } = require("uuid");
 const slugify = require("slugify");
-const { TUTORIAL_STATUS } = require("../../../shared/tutorialStatus");
+const { TUTORIAL_STATUS } = require("../../../../../shared/tutorialStatus");
 
 exports.createTutorial = async (data, trx = db) => {
   const insertData = { included_plans: [], ...data };
