--- backend/src/shared/tutorialStatus.js
@@ -1,7 +1,0 @@
-const TUTORIAL_STATUS = Object.freeze({
-  DRAFT: 'draft',
-  PUBLISHED: 'published',
-  ARCHIVED: 'archived',
-});
-
-module.exports = { TUTORIAL_STATUS };
