+++ frontend/src/pages/dashboard/admin/tutorials/index.js
@@ -26,7 +26,7 @@ import { sendChatMessage } from "@/services/messageService";
 import useAuthStore from "@/store/auth/authStore";
 import useNotificationStore from "@/store/notifications/notificationStore";
 import useMessageStore from "@/store/messages/messageStore";
-import { TUTORIAL_STATUS } from "@/shared/tutorialStatus";
+import { TUTORIAL_STATUS } from "@shared/tutorialStatus";
 
 function AdminTutorialsPage() {
   const { t } = useTranslation("dashboard", { keyPrefix: "tutorialsPage" });
