--- frontend/src/shared/socialPlatforms.json
@@ -1,9 +1,0 @@
-[
-  "linkedin",
-  "github",
-  "twitter",
-  "youtube",
-  "facebook",
-  "instagram",
-  "website"
-]
