--- frontend/src/shared/tutorialStatus.js
@@ -1,7 +1,0 @@
-export const TUTORIAL_STATUS = Object.freeze({
-  DRAFT: "draft",
-  PUBLISHED: "published",
-  ARCHIVED: "archived",
-});
-
-export default TUTORIAL_STATUS;
